qemuboot.bbclass: Handle the QB_DEFAULT_KERNEL value fallback
The QB_DEFAULT_KERNEL value represents the desired kernel image type the
QEMU machine expects to load. However because QB_DEFAULT_KERNEL is
rewritten in qemuboot.bbclass its original value can no longer be easily
reversed from the full file path (the target of the symlink).

When runqemu tries to find a kernel image it will first try the
QB_DEFAULT_KERNEL however will never use the original value in
QB_DEFAULT_KERNEL (as it is not provided in the .conf) instead it will
fallback to the KERNEL_IMAGETYPE value. To allow runqemu to search for
the original value of QB_DEFAULT_KERNEL, rewrite KERNEL_IMAGETYPE to the
value of QB_DEFAULT_KERNEL within the output .conf.

This use case is important for where KERNEL_IMAGETYPES has multiple
values or for example when the QEMU machine needs to be loaded with a
"*-initramfs" image.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>