ghostscript: Define COMPATIBLE_HOST with supported platforms
Setup the COMPATIBLE_HOST variable with all supported hosts based on the
provided objarch.h files. This inverts the existing pattern where
unsupported hosts (e.g. 'arc.*') are explicitly set.

This suppresses the warning message generated on hosts that are not
supported but were not listed explicitly.

  WARNING: .../ghostscript_9.26.bb: Unable to get checksum for
  ghostscript SRC_URI entry objarch.h: file could not be found

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>
---
Changes in v4:
  - Update for 9.50 recipe
Changes in v3:
  - Update for 9.26 recipe
Changes in v2:
  - Fix up to use HOST_SYS patterns instead of OVERRIDES
  - Make i586/i686 pattern common as done in other recipes