WIP: Add gnu testsuite execution for OEQA
This patch is an RFC for adding support to execute the gnu test suites
for binutils, gcc and glibc. With the intention for enabling automated
test running of these test suites within the OEQA framework such that
they can be executed by the Yocto Autobuilder.

Please note that this patch is a work in progress. For ease of reference
and review the entire changeset is currently provided as a single commit
including work in progress patches for gcc.

The test suites covered need significant resources or build artifacts
such that running them on the target is undesirable which rules out the
use of ptest. Because of this the test suites can be run on the build
host and call out to the target for execution (via ssh or using qemu
usermode).

The following implementation adds a do_check task to binutils-cross,
gcc-cross and gcc-runtime in order to execute the test suite on the
build host. For glibc, a second recipe is created which includes the
base glibc recipe. The reason for this is due to the libgcc -> glibc
-> libgcc-initial dependency chain and the requirements of the test
suite to have libgcc for compilation and execution testing.

Target execution is another important issue specifically since target
execution is slow to extremely slow depending on the physical or
emulated target performance. In order to provide faster execution
performance qemu linux-user is implemented alongside qemu system
emulation (via ssh+nfs). In initial testing qemu linux user vs qemu
system emulation provided performance gains of between 10x to 60x whilst
initially having small pass/fail differences. Further work as covered in
this patch reduces the pass/fail differences for gcc/gcc-runtime to
almost 0 (with filtering of the small amount of known qemu user
failures).

However glibc is more strict with its expectations of CPU implementation
as well as syscall behaviour. Additionally glibc expects to be able to
execute OS tools such as 'sh' and 'echo' of which is not easy to provide
in the recipe-sysroot (bindir is not populated into the sysroot). As
such correct test results for glibc rely on execution with qemu system
emulation or on a physical target. This patch however still includes
qemu user execution for glibc with a large set of filtered tests that
are known to fail specifically under qemu user execution.

(failed tests / total tests / filtered test failures)
           | binutils      | gas           | gold         | ld           | libiberty
aarch64    |    0/  201/ 0 |    0/  443/ 0 |   0/    9/ 0 |   0/ 1601/ 0 |   0/   28/ 0
arm        |    0/  202/ 0 |    0/  873/ 0 |   0/   26/ 0 |   0/ 1603/ 1 |   0/   28/ 0
i686       |    0/  207/ 0 |    0/ 1257/ 0 |   0/    6/ 0 |   0/ 1715/ 0 |   0/   28/ 0
mips       |    0/  225/ 0 |    0/ 7222/ 0 |   0/    5/ 1 |   0/ 1804/21 |   0/   28/ 0
mips64     |    0/  224/ 0 |    0/ 7246/ 0 |   0/    5/ 1 | 646/ 1750/15 |   0/   28/ 0
powerpc    |    0/  198/ 0 |    0/  266/ 1 |   0/    5/ 0 |   0/ 1503/ 0 |   0/   28/ 0
riscv64    |    0/  196/ 0 |    0/  258/ 0 |              |   0/ 1150/ 4 |   0/   28/ 0
x86_64     |    0/  247/ 0 |    0/ 1257/ 0 |   0/   11/ 0 |   0/ 2138/ 0 |   0/   28/ 0

           | gcc           | g++           | libatomic    | libgomp      | libitm       | libstdc++-v3
aarch64    |   6/130911/ 1 |   0/129010/ 0 |   0/   54/ 0 |   0/ 2515/ 0 |   0/   46/ 0 |   1/12787/ 0
arm        |   0/122468/ 3 |   1/128403/ 0 |   0/   49/ 0 |   0/ 2515/ 0 |   0/   46/ 0 |   0/12786/19
i686       |   1/133589/39 |   0/130255/ 0 |   0/   49/ 0 |   0/ 2522/ 0 |   0/   46/ 0 |   1/13006/17
mips       | 117/132171/ 1 |  20/128150/ 0 |   4/   49/ 0 |   2/ 2515/ 0 |              |  14/12778/17
mips64     | 169/134736/ 5 |  49/130221/ 0 |   1/   54/ 0 |   0/ 2515/ 0 |              |   8/12778/ 0
powerpc    | 341/119777/ 5 |  19/129082/ 0 |   0/   49/ 0 |   2/ 2515/ 0 |   0/   46/ 0 |  16/12994/17
riscv64    |  10/106406/ 2 |  24/128453/ 0 |   0/   54/ 0 |   4/ 2512/ 0 |              |   1/12746/ 0
x86_64     |   0/135180/24 |   0/131939/ 0 |   0/   54/ 0 |   0/ 2522/ 0 |   0/   46/ 0 |   2/13006/ 0
x86_64-kvm |   3/135420/ 4 |   0/131939/ 0 |   0/   54/ 0 |   1/ 2522/ 0 |  18/   46/ 0 |  54/13008/ 0

           | glibc         | glibc-2.30
aarch64    |   56/ 5882/19 |   58/ 5994/19
arm        |   49/ 5130/18 |   51/ 5240/19
i686       | 1464/ 5900/ 9 | 1466/ 6010/10
mips       |   63/ 5094/19 |   63/ 5199/19
mips64     | 2014/ 5847/19 | 2017/ 5952/18
powerpc    | 1202/ 5110/19 | 1207/ 5254/19
riscv64    |   68/ 5847/19 |   68/ 5952/18
x86_64     | 1417/ 5991/38 | 1421/ 6101/39
x86_64-kvm |   26/ 5982/ 0 |   46/ 6092/ 0

This patch also introduces some OEQA test cases which cover running the
test suites. The test cases are split into binutils, gcc and glibc.
Individual test cases provide execution of the sub-suites of tests
within each target. For example binutils has binutils, gas, gold, ld and
libiberty suites which can each be executed independently.

The OEQA test cases implement execution on qemu linux user by default.
Subclasses implement qemu system emulation setup and configuration for
running the tests.

Changes in v2:
- Moved gcc/binutils test suite recipes into tasks on
  binutils-cross/gcc-cross/gcc-runtime
- Split selftest module into separate classes for binutils/gcc/glibc
- Split selftest test cases into associated testsuites
  (e.g. gcc -> gcc, g++)
- Added test result parsing into pass/fail/etc.
- Added system emulation test cases into selftest module (including unfs
  setup)
- Added test case filtering and asserting failure to selftest module
- Fixes and improvements to resolve test case failures (primarily for
  binutils/gcc)
- GCC patches to mark specific warnings as acceptable output
- GCC patches to fix specific tests
- Fix for OE specific binutils patch
- Testing against glibc 2.30
---
RFC: 1
Cc: Manjukumar Harthikote Matha <manjukum@xilinx.com>
Cc: Khem Raj <raj.khem@gmail.com>
Cc: Richard Purdie <richard.purdie@linuxfoundation.org>
Cc: Ross Burton <ross.burton@intel.com>